Separated route functions from server.js into their own files under controllers dir.
