+++ controllers/image.js
@@ -1,0 +1,15 @@
+// Handle image route behavior
+const handleImagePut = (req, res, db) => {
+    const { id } = req.body;
+    db('users').where('id', '=', id)
+        .increment('entries', 1)
+        .returning('entries')
+        .then(entries => {
+            res.json(entries[0]);
+        })
+        .catch(err => res.status(400).json('Unable to get entries'));
+}
+
+module.exports = {
+    handleImagePut: handleImagePut
+}
