+++ controllers/signin.js
@@ -1,0 +1,24 @@
+// Handle signin route behavior
+const handleSignIn = (req, res, db, bcrypt) => {
+    db.select('email', 'hash').from('login')
+        .where('email', '=', req.body.email)
+        .then(data => {
+            // bcrypt compare password and hash values
+            const isValid = bcrypt.compareSync(req.body.password, data[0].hash);
+            if (isValid) {
+                return db.select('*').from('users')
+                    .where('email', '=', req.body.email)
+                    .then(user => {
+                        res.json(user[0]);
+                    })
+                    .catch(err => res.status(400).json('Unable to get user'));
+            } else {
+                res.status(400).json('Wrong credentials');
+            }
+        })
+        .catch(err => res.status(400).json('Wrong credentials'));
+}
+
+module.exports = {
+    handleSignIn: handleSignIn
+}
