+++ server.js
@@ -2,7 +2,12 @@ const express = require('express');
 const bcrypt = require('bcrypt');
 const cors = require('cors');
 const knex = require('knex');
-const { response } = require('express');
+
+// Import controllers for routes
+const register = require('./controllers/register');
+const signin = require('./controllers/signin');
+const profile = require('./controllers/profile');
+const image = require('./controllers/image');
 
 // Connect to database
 const db = knex({
@@ -24,95 +29,11 @@ app.use(express.json());
 app.use(cors());
 
 //================================== ROUTES ====================================
-
-// ROOT
-app.get('/', (req, res) => {
-    res.send('Success');
-});
-
-// SIGN IN
-app.post('/signin', (req, res) => {
-    db.select('email', 'hash').from('login')
-        .where('email', '=', req.body.email)
-        .then(data => {
-            // bcrypt compare password and hash values
-            const isValid = bcrypt.compareSync(req.body.password, data[0].hash);
-            if (isValid) {
-                return db.select('*').from('users')
-                    .where('email', '=', req.body.email)
-                    .then(user => {
-                        res.json(user[0]);
-                    })
-                    .catch(err => res.status(400).json('Unable to get user'));
-            } else {
-                res.status(400).json('Wrong credentials');
-            }
-        })
-        .catch(err => res.status(400).json('Wrong credentials'));
-});
-
-// REGISTER
-app.post('/register', (req, res) => {
-    const { email, name, password } = req.body;
-
-    // bcrypt hash password string
-    const saltRounds = 10; // bcrypt cost factor
-    const hash = bcrypt.hashSync(password, saltRounds);
-
-    // Create a transaction (2 actions)
-    db.transaction(trx => {
-        trx.insert({
-            hash: hash,
-            email: email
-        })
-        .into('login')
-        .returning('email')
-        .then(loginEmail => {
-            // Insert user info into db users table
-            return trx('users')
-                .returning('*')
-                .insert({
-                    email: loginEmail[0],
-                    name: name,
-                    joined: new Date()
-                })
-                .then(user => {
-                    // respond with user json data
-                    res.json(user[0]);
-                })
-        })
-        .then(trx.commit)
-        .catch(trx.rollback);
-    })
-    .catch(err => res.status(400).json('Unable to register'));
-});
-
-// PROFILE
-app.get('/profile/:id', (req, res) => {
-    const { id } = req.params;
-    db.select('*').from('users').where({
-        id: id
-    }).then(user => {
-        if (user.length) {
-            res.json(user[0]);
-        } else {
-            res.status(400).json('Not found');
-        }
-    })
-    .catch(err => res.status(400).json('Error getting user'));
-});
-
-// IMAGE
-app.put('/image', (req, res) => {
-    const { id } = req.body;
-    db('users').where('id', '=', id)
-        .increment('entries', 1)
-        .returning('entries')
-        .then(entries => {
-            res.json(entries[0]);
-        })
-        .catch(err => res.status(400).json('Unable to get entries'));
-});
+app.get('/', (req, res) => { res.send('Success') });
+app.post('/signin', (req, res) => { signin.handleSignIn(req, res, db, bcrypt)});
+app.post('/register', (req, res) => { register.handleRegister(req, res, db, bcrypt) });
+app.get('/profile/:id', (req, res) => { profile.handleProfileGet(req, res, db) });
+app.put('/image', (req, res) => { image.handleImagePut(req, res, db) });
 
 // Start Server
 app.listen(3001, () => {
